requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
class==1.0.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
et_xmlfile==2.0.0
filelock==3.16.1
finance-datareader==0.9.94
fonttools==4.55.3
fpdf==1.7.2
fpdf2==2.8.2
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaleido==0.2.1
kiwisolver==1.4.8
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
multidict==6.1.0
narwhals==1.22.0
numpy==2.2.1
openai==1.59.7
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
propcache==0.2.1
protobuf==5.29.3
pyarrow==18.1.0
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.28.0
webencodings==0.5.1
wheel==0.45.1
xlrd==2.0.1
yarl==1.18.3
yt-dlp==2024.12.23
